requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
Django==3.2.10
django-environ==0.11.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
idna==3.6
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
whitenoise==6.6.0 #must be in requirements.txt
gunicorn==21.2.0 #must be in requirements.txt